Fixed type serialization for meetup job.

+++ server/jobmine_jobs/remind_meetup_job/remind_meetup_job.go
@@ -56,7 +56,7 @@ func packageTaskRecordMetadata(
 }
 
 func parseUserInfo(taskRecord jobmine.TaskRecord) (data.TUserID, UserType) {
-	userId := data.TUserID(taskRecord.Metadata[USER_ID_METADATA_KEY].(uint))
+	userId := data.TUserID(uint(taskRecord.Metadata[USER_ID_METADATA_KEY].(float64)))
 	userType := UserType(taskRecord.Metadata[USER_TYPE_METADATA_KEY].(string))
 	return userId, userType
 }
